# Conflicts:
#	BDS/ServiceInterface/WebEas.Esam.ServiceInterface.Office/EsamAppHostBase.cs
#	BDS/ServiceInterface/WebEas.Esam.ServiceInterface.Office/RepositoryBase.cs
#	BDS/ServiceInterface/WebEas.Esam.ServiceInterface.Pfe/PfeRepository.cs
#	BDS/ServiceInterface/WebEas.Esam.ServiceInterface.Pfe/PfeService.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/DataChangesRequest.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/OrsElementPermissionDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/OrsElementTypeDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/RightPermissionDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/RoleDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/TenantDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/TreeDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/TreePermissionDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Dto/UserDto.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Types/Tenant.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Types/Tree.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office/Combo.cs
#	BDS/ServiceModel/WebEas.Esam.ServiceModel.Office/WebEas.Esam.ServiceModel.Office.csproj
#	BDS/Web/Office/WebEas.Esam.Office.Cfe/Web.config
#	BDS/Web/WebEas.Esam.Pfe/AppHost.cs
#	BDS/Web/WebEas.Esam.Pfe/Properties/PublishProfiles/DevelopProfile.pubxml
#	BDS/Web/WebEas.Esam.Pfe/Web.config

+++ BDS/ServiceModel/WebEas.Esam.ServiceModel.Office.Cfe/Types/TenantType.cs
@@ -12,7 +12,7 @@ namespace WebEas.Esam.ServiceModel.Office.Cfe.Types
         [PrimaryKey]
         [AutoIncrement]
         [DataMember]        
-        public short C_TenantType_Id { get; set; }
+        public byte C_TenantType_Id { get; set; }
 
         [DataMember]
         [PfeColumn(Text = "Typ", Mandatory = true)]
